(MoritzLangenstein) Refactored curation and institutions
