+++ iaso/institutions/__init__.py
@@ -7,8 +7,8 @@ from itertools import zip_longest
 from tqdm import tqdm
 
 from .wikidataclient import WikiDataClient
-from .institution_entity_extraction import greedily_extract_institution_entities
-from .institution_entity_linking import query_institution_entity_details
+from .entity.matching import greedily_match_institution_entities
+from .entity.linking import query_institution_entity_details
 
 DETAILS_CHUNK_SIZE = 10
 
@@ -27,9 +27,9 @@ async def deduplicate_registry_institutions(registry, academine_path):
     async def collect_extracted_institution_entities(
         client, institution_string, iid, progress, extracted_institution_entities
     ):
-        extracted_institution_entities[
-            iid
-        ] = await greedily_extract_institution_entities(client, institution_string)
+        extracted_institution_entities[iid] = await greedily_match_institution_entities(
+            client, institution_string
+        )
 
         progress.update()
         progress.set_postfix(ordered_dict=client.backoff)
